+++ wow_world_messages/CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * `From` conversion for update mask builder types to proper types.
 * `From` conversion for update mask types to `UpdateMask` type.
 * `Guid::from_u32s` and `Guid::to_u32s` to create a Guid from high/low `u32`s.
+* `Octal`, `Binary`, `UpperHex`, `LowerHex` for `Guid`.
 * `std::fmt::Display` for new enums.
 
 ### Changed